+++ pinot-core/src/main/java/org/apache/pinot/core/segment/processing/genericrow/AdaptiveConstraintsWriter.java
@@ -16,34 +16,12 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.core.segment.processing.framework;
+package org.apache.pinot.core.segment.processing.genericrow;
 
-public class AdaptiveSizeBasedConstraintsChecker implements AdaptiveConstrainsChecker {
+import java.io.IOException;
 
-  private final long _bytesLimit;
-  private long _numBytesWritten = 0;
 
-  public AdaptiveSizeBasedConstraintsChecker(long bytesLimit) {
-    _bytesLimit = bytesLimit;
-  }
-
-  public boolean canWrite() {
-    return _numBytesWritten < _bytesLimit;
-  }
-
-  public long getNumBytesWritten() {
-    return _numBytesWritten;
-  }
-
-  public long getBytesLimit() {
-    return _bytesLimit;
-  }
-
-  public void updateNumBytesWritten(long numBytesWritten) {
-    _numBytesWritten += numBytesWritten;
-  }
-
-  public void reset() {
-    _numBytesWritten = 0;
-  }
+public interface AdaptiveConstraintsWriter<W, D> {
+  boolean canWrite();
+  void write(W writer, D dataUnit) throws IOException;
 }
